Moved view files to the appropriate folder.
