+++ src/main/java/io/display_logic/ViewPresenter.java
@@ -1,4 +1,4 @@
-package CoreEntities.IO;
+package io.display_logic;
 
 public class ViewPresenter {
 }
